Split test: `pkg/db/postgres/data`
Because of the size of the test file
